+++ lib/doorkeeper/models/access_grant_mixin.rb
@@ -9,6 +9,7 @@ module Doorkeeper
     include Models::Revocable
     include Models::Accessible
     include Models::Orderable
+    include Models::Hashable
     include Models::Scopes
 
     # never uses pkce, if pkce migrations were not generated
@@ -30,7 +31,7 @@ module Doorkeeper
       #   if there is no record with such token
       #
       def by_token(token)
-        find_by(token: token.to_s)
+        find_by_hashed_or_plain_token(:token, token)
       end
 
       # Revokes AccessGrant records that have not been revoked and associated