+++ spec/lib/config_spec.rb
@@ -525,4 +525,17 @@ describe Doorkeeper, 'configuration' do
       expect(subject.handle_auth_errors).to eq(:raise)
     end
   end
+
+  describe 'hash_secrets together with reuse_access_token' do
+    it 'will disable reuse_access_token' do
+      expect(Rails.logger).to receive(:warn).with(/reuse_access_token will be disabled/)
+
+      Doorkeeper.configure do
+        reuse_access_token
+        hash_secrets { 'does not matter' }
+      end
+
+      expect(subject.reuse_access_token).to eq(false)
+    end
+  end
 end